Rename MathUtils to o.l.util.math.Scalars
