+++ lenskit-core/src/main/java/org/grouplens/lenskit/vectors/similarity/CosineVectorSimilarity.java
@@ -22,8 +22,8 @@ package org.grouplens.lenskit.vectors.similarity;
 
 import com.google.common.base.Preconditions;
 import org.grouplens.lenskit.core.Shareable;
-import org.grouplens.lenskit.util.MathUtils;
 import org.grouplens.lenskit.vectors.SparseVector;
+import org.lenskit.util.math.Scalars;
 
 import javax.inject.Inject;
 import java.io.Serializable;
@@ -62,7 +62,7 @@ public class CosineVectorSimilarity implements VectorSimilarity, Serializable {
     public double similarity(SparseVector vec1, SparseVector vec2) {
         final double dot = vec1.dot(vec2);
         final double denom = vec1.norm() * vec2.norm() + dampingFactor;
-        if (MathUtils.isZero(denom)) {
+        if (Scalars.isZero(denom)) {
             return 0;
         } else {
             return dot / denom;
