+++ lenskit-core/src/main/java/org/grouplens/lenskit/vectors/similarity/DistanceVectorSimilarity.java
@@ -23,9 +23,9 @@ package org.grouplens.lenskit.vectors.similarity;
 import it.unimi.dsi.fastutil.longs.LongSet;
 import org.grouplens.lenskit.collections.LongUtils;
 import org.grouplens.lenskit.core.Shareable;
-import org.grouplens.lenskit.util.MathUtils;
 import org.grouplens.lenskit.vectors.MutableSparseVector;
 import org.grouplens.lenskit.vectors.SparseVector;
+import org.lenskit.util.math.Scalars;
 
 import javax.inject.Inject;
 import java.io.Serializable;
@@ -53,7 +53,7 @@ public class DistanceVectorSimilarity implements VectorSimilarity, Serializable 
     public double similarity(SparseVector vec1, SparseVector vec2) {
         final double distance;
         // One of the vector is empty
-        if (MathUtils.isZero(vec1.norm()) || MathUtils.isZero(vec2.norm())){
+        if (Scalars.isZero(vec1.norm()) || Scalars.isZero(vec2.norm())){
             return Double.NaN;
         }
