+++ lenskit-core/src/main/java/org/grouplens/lenskit/vectors/similarity/SpearmanRankCorrelation.java
@@ -24,9 +24,9 @@ import com.google.common.primitives.Doubles;
 import it.unimi.dsi.fastutil.longs.AbstractLongComparator;
 import it.unimi.dsi.fastutil.longs.LongArrays;
 import org.grouplens.lenskit.core.Shareable;
-import org.grouplens.lenskit.util.MathUtils;
 import org.grouplens.lenskit.vectors.MutableSparseVector;
 import org.grouplens.lenskit.vectors.SparseVector;
+import org.lenskit.util.math.Scalars;
 
 import javax.inject.Inject;
 import java.io.Serializable;
@@ -76,7 +76,7 @@ public class SpearmanRankCorrelation implements VectorSimilarity, Serializable {
             int j;
             for (j = i + 1; j < n; j++) {
                 // compare difference to 0 with tolerance - more robust
-                if (!MathUtils.isZero(values[j] - values[i])) {
+                if (!Scalars.isZero(values[j] - values[i])) {
                     break;
                 }
             }
